move `RecordingWebServer` to `tests.util.misc` (#17480)
