+++ tests/conftest.py
@@ -69,7 +69,7 @@ from chia.wallet.wallet_node_api import WalletNodeAPI
 from tests.core.data_layer.util import ChiaRoot
 from tests.core.node_height import node_height_at_least
 from tests.simulation.test_simulation import test_constants_modified
-from tests.util.misc import BenchmarkRunner, GcMode, _AssertRuntime, measure_overhead
+from tests.util.misc import BenchmarkRunner, GcMode, RecordingWebServer, _AssertRuntime, measure_overhead
 from tests.util.setup_nodes import (
     OldSimulatorsAndWallets,
     SimulatorsAndWallets,
@@ -1175,3 +1175,15 @@ async def farmer_harvester_2_simulators_zero_bits_plot_filter(
         )
 
         yield farmer_service, harvester_service, simulators[0], simulators[1], bt
+
+
+@pytest.fixture(name="recording_web_server")
+async def recording_web_server_fixture(self_hostname: str) -> AsyncIterator[RecordingWebServer]:
+    server = await RecordingWebServer.create(
+        hostname=self_hostname,
+        port=uint16(0),
+    )
+    try:
+        yield server
+    finally:
+        await server.await_closed()
